xmltream.py: rework SSL/TLS implementation.
Due to historical issues, the original implementation of SSL/TLS
in xmlstream.py hardcoded self.ssl_version = ssl.PROTOCOL_TLSv1,
which is the only supported TLS in Python at the time, and causes
a interoperability problem , that prevents connecting to a server with
TLSv1.1/1.2 but doesn't support TLSv1.

Simply changing ssl_version to ssl.PROTOCOL_SSLv23, that allows all
versions of SSL/TLS, including TLSv1.2 in newer Python, could fix the
problem. However, it introduces more security hazards, since SSLv2
and SSLv3 is not disabled, the connection is vulnerable to various
attacks.

In addition, the implementation did not use the best practice when
establishing connections, for example, any ciphers provided by OpenSSL
is used, including RC4, which is vulnerable; SSL compression may be
used, which is vulnerable to CRIME attack.

Therefore, a rework of the entire SSL/TLS implementation is needed,
to fix the interoperability problem, and provide better connections.

This commit introduces the following changes:

* Default ssl_version is set as ``PROTOCOL_SSLv23``, but SSLv2
and SSLv3 is explicitly disabled, allowing us to use any latest TLS
versions. As a fallback, if users really have to work with an insecure
server, ``PROTOCOL_SSLv3`` may be set.

* The most complicated part, is the new behavior on creating SSL
sockets. Even if new features is provided in stdlib, we have to make it
it secure on different versions of Python(s).

* First, if Python 3.4/2.7.9 or newer is found, we are going to use
ssl.create_default_context, a new function, that provides (partially)
secure settings. This is the best choice, allowing security updates
to be delivered with Python.

* Second, if Python 3 is used, but not as new as 3.4, we use
ssl.SSLContext(), and initialize it with ssl.PROTOCOL_TLSv1, because
this is the only supported TLS in previous Python. We also set a list
of good ciphers, _CIPHERS_TLS, which is taken from latest Python. Then
we disable SSL Compression to prevent CRIME attacks.

* Third, if Python 2.7 is running, but not as new as 2.7.9, we have
to call ssl.wrap_socket() directly, and use ssl.PROTOCOL_TLSv1.

* As the fallback mentioned before, a user may have to deal with a
server with outdated and insecure SSLv3, we need to add other conditions
to make sure a SSLv3 connection can be established. To achieve this,
if ssl.PROTOCOL_SSLv3 is used, for Python 3.4/2.7.9, we deselect
ssl.OP_NO_SSLv3, or select SSLv3; for older Python 3 or 2.7, we
initialize with ssl.PROTOCOL_SSLv3 instead.

* This is not enough, our ciphers list should include ciphers that SSLv3
servers can understand. So, a list of ciphers, _CIPHERS_SSL, which is
also taken from Python stdlib, but with a 3DES fallback.

* Finally, still due to historical problem, early Python can not verify
server-side certificates, unless a file contains CAs is specified. But in
newer Python, a system-wide CAs can be used for verification without specifying
CAs manually. But in xmltream.py, the presence of self.ca_certs,
directly disables/enables the verification. In the new code, this
original behavior is respected and implemented.

After all, this commit strengthen the SSL/TLS security, but does not
complete solves it. In other words, in most circumstances, server-side
certificates are not verified at all. A man in the middle just needs to
send a self-signed certificate to hijack your connection, making all the
changes mentioned before meaningless. We have to figure out a way to add
a new option to fix this, but also compatible with current library
users.

Signed-off-by: Tom Li <biergaizi@member.fsf.org>